requirements.txt
Django>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
celery>=5.2
redis>=4.0
psycopg2-binary>=2.9
skyfield>=1.45
